requirements.txt
flask
flask-cors
psycopg2-binary
openai
tavily-python
gunicorn